==> 01.1-DATA-web_scrapper-try-except.ipynb <==
# Libraries

import pandas as pd

import random
import time
import requests

from bs4 import BeautifulSoup

# Scrapping the website

news = []

%%time
for i in range (5100, 5250):
    url = "https://www.meneame.net/" + str(i)

    header = {
        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36",
        "accept-language": "en-GB,en-US;q=0.9,en;q=0.8",
    }

    
    seg = random.randint(5,15)
    time.sleep(seg)
    
    web = requests.get(url, headers=header)
    print(web)
    
    soup = BeautifulSoup(web.text, "html.parser")
    part_news = soup.find_all("div", attrs={"class":"link"})
    
    print(f'{url} done with {len(part_news)} entries')
    
    news.extend(part_news)


# Creating a Dataframe

### Checking BeautifulSoup commands

title = news[0].find('h2').text
title

body = news[0].find(class_='link-preview').text
body

category = news[0].find(class_='community-name').span.text
category

votes = news[0].find(class_='votes-counter').text
votes

clicks = news[0].select('div.left-actions a span.comments-number')[0].text
clicks

comments = news[0].select('div.left-actions a span.comments-number')[1].text
comments

### Loop through the results

news_data = {'title':[], 'body':[], 'categroy':[], 'votes':[], 'clicks':[], 'comments':[]}
error_counter = 0
for i in news:
    try:
        title = i.find('h2').text
        body = i.find(class_='link-preview').text
        category = i.find(class_='community-name').span.text
        votes = i.find(class_='votes-counter').text
        clicks = i.find(class_='comments-number').text
        comments = i.select('div.left-actions a span.comments-number')[1].text
        
        news_data['title'].append(title)
        news_data['body'].append(body)
        news_data['category'].append(topic)
        news_data['votes'].append(votes)
        news_data['clicks'].append(clicks)
        news_data['comments'].append(comments)

    except:
        error_counter += 1


error_counter

news_df_raw = pd.DataFrame(news_data)

# Save raw data

news_df_raw.to_csv('../DATA/news_125K_raw.csv')
