==> 01.2-DATA-cleaning_raw_data.ipynb <==
# Libraries

import pandas as pd

# Import Data

news_df = pd.read_csv('../DATA/news_125K_raw.csv', index_col=0)

news_df

# Data cleaning

### Removing

#### Remove unwanted columns

news_df.drop(columns=['Unnamed: 0.1', 'Unnamed: 0'], axis=1, inplace=True)

#### Remove duplicated rows

news_df.duplicated().sum()

news_df.drop_duplicates(inplace=True)

#### Remove subset duplicated rows

news_df.describe(include='all')

news_df.drop_duplicates(subset='title', inplace=True)

news_df.drop_duplicates(subset='body', inplace=True)

### Topic feature

#### Regroup topic classification

news_df['topic'].value_counts()[0:227]

news_df.loc[news_df['topic']=='Únete a Actualidad','topic'] = 'Latest'
news_df.loc[news_df['topic']=='Únete a Cultura','topic'] = 'Culture'
news_df.loc[news_df['topic']=='Únete a Tecnología','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Ocio','topic'] = 'Entertainment'
news_df.loc[news_df['topic']=='Únete a Politica','topic'] = 'Politics'
news_df.loc[news_df['topic']=='Únete a Ciencia','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Astronomia','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Linux','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Vacúname','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Sucesos','topic'] = 'Latest'
news_df.loc[news_df['topic']=='Únete a Videojuegos','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Motor','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Barcos','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Podemos','topic'] = 'Politics'
news_df.loc[news_df['topic']=='Únete a Softlibre','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Sysdevs','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Trenes','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Series','topic'] = 'Culture'
news_df.loc[news_df['topic']=='Únete a Música','topic'] = 'Culture'
news_df.loc[news_df['topic']=='Únete a Coronavirus','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Salud','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Cine','topic'] = 'Culture'
news_df.loc[news_df['topic']=='Únete a Botánica','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Idiomas','topic'] = 'Culture'
news_df.loc[news_df['topic']=='Únete a Mambiente','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Sequía','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Anarquismo','topic'] = 'Politics'
news_df.loc[news_df['topic']=='Únete a Motociclismo','topic'] = 'Technology'
news_df.loc[news_df['topic']=='Únete a Fotografía','topic'] = 'Culture'
news_df.loc[news_df['topic']=='Únete a Medicina','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Biologia','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Música70','topic'] = 'Culture'
news_df.loc[news_df['topic']=='Únete a Psicología','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Feminismo','topic'] = 'Politics'
news_df.loc[news_df['topic']=='Únete a Energias','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Neurociencia','topic'] = 'Science'
news_df.loc[news_df['topic']=='Únete a Mnm','topic'] = 'Latest'


#### Assign 'Uncategorized' topic to the rest

topics = ['Latest', 'Culture', 'Technology', 'Entertainment', 'Politics', 'Science']
news_df['topic'][~news_df['topic'].isin(topics)]  = 'Uncategorized'


news_df['topic'].value_counts()

### Clicks, votes and comments fetaure

#### Substitute the 'k' for 00 (thousands)

news_df['clicks'] = news_df['clicks'].str.replace('k', '00', regex=True)
news_df['votes'] = news_df['votes'].str.replace('k', '00', regex=True)
news_df['comments'] = news_df['comments'].str.replace('k', '00', regex=True)

#### Delete any character that is not a decimal digit

news_df['clicks'] = news_df['clicks'].str.replace(r'\D+', '', regex=True)
news_df['votes'] = news_df['votes'].str.replace(r'\D+', '', regex=True)
news_df['comments'] = news_df['comments'].str.replace(r'\D+', '', regex=True)

### Nans or null values

news_df.isna().sum()

#### Substitute Nans in comments for 0

news_df['comments'].fillna(0, inplace=True)

#### Delete empty body rows

news_df.dropna(inplace=True)

### Dtypes

news_df.info()

dtypes = {'votes': int,
         'clicks': int,
         'comments': int}

news_df = news_df.astype(dtypes)

# Checking

news_df.describe(include='all')

news_df.info()

df

- There are no duplicated rows
- There are no null or nan values
- Values are clear and meaningfull
- Dtypes are coherent
- Data is consistent and robust

Other issues will be handled in further developments

# Save file

news_df.to_csv('../DATA/news_125k.csv')
