==> 02.3-NLP-sentiment_analysis.ipynb <==
# Libraries

import pandas as pd 
import numpy as np 
import re 

import matplotlib.pyplot as plt 
import seaborn as sns 
%matplotlib inline

# Constants

DATA_PATH = '../DATA/news_125k-tokens-relevant.csv'

# Bespoke Functions

# Import data

df = pd.read_csv(DATA_PATH, index_col=0)
df.head()

# Sentiment Anlysis

!pip install transformers

df_sentiment_try = df.iloc[:]

text_polarity = []
text_score = []

### Create model

from transformers import pipeline

classifier = pipeline('sentiment-analysis', 
                      model="nlptown/bert-base-multilingual-uncased-sentiment")

%%time
for text in df_sentiment_try['clean_text']:    
    if len(text) > 512:
        text = text[:512] # max of 512, otherwise, gives error.
    else:
        pass
    results = classifier(text)
    
    # get the polarities
    polarity = results[0]['label']
    text_polarity.append(polarity)
    

    # get the score
    score = round(results[0]['score'], 2)
    text_score.append(score)

### Add results to the dataframe

df_sentiment_try['polarity'] = text_polarity
df_sentiment_try['polarity_score'] = text_score

len(text_polarity), len(text_score)

# Save file

df_sentiment_try.to_csv('../DATA/news_125k-tokens-relevant-with_polarity.csv')

### Using TextBlob

!pip install -U textblob

!python -m textblob.download_corpora

from textblob import TextBlob

textblob_polarity = []
textblob_subjectivity = []


for text in df['clean_text']:
    # first we need to translate 
    text_blob = TextBlob(text)
    text_blob.translate(to='en')
    
    # get the polarities
    text_polarity = text_blob.sentiment.polarity
    textblob_polarity.append(text_polarity)
    
    # get the subjectivities
    text_subjectivity = text_blob.sentiment.subjectivity
    textblob_subjectivity.append(text_subjectivity)

df['polarity'] = textblob_polarity
df['subjectivity'] = textblob_subjectivity
