==> 03.2-EDA-analysis.ipynb <==
# Libraries

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
plt.style.use('ggplot')

# Constants

DATA_PATH = '../DATA/news_125k-tokens-relevant.csv'

# Bespoke Functions

# Import data

df = pd.read_csv(DATA_PATH, index_col=0)
df

df.info()

### Import Topic

df_topic = pd.read_csv('../DATA/news_125k-tokens-relevant-with_topic.csv', index_col=0)
df_topic

df_topic.info()

df_topic.isna().sum()

### Import Polarity

df_polarity = pd.read_csv('../DATA/news_125k-tokens-relevant-with_polarity.csv', index_col=0)
df_polarity

df_polarity.info()

df_polarity.isna().sum()

### Import Score

df_score = pd.read_csv('../DATA/news_125k-with_score.csv', index_col=0)
df_score

df_score.info()

df_score.iloc[1341]

### Add polarity, topic and score columns to df

df['topic'] = df_topic['topic']
df['topic_name'] = df_topic['topic_name']

df['polarity'] = df_polarity['polarity']
df['polarity_score'] = df_polarity['polarity_score']

df['score'] = df_score['score']
df['score_name'] = df_score['score_name']
df['rank'] = df_score['rank']
df['status'] = df_score['status']

df

df.info()

# Clean data

df['polarity'] = df['polarity'].str.replace(r'\D+', '', regex=True)
df['polarity'].value_counts()

df['polarity'] = df['polarity'].astype('int')


df.info()

### Number of relevant

df['num_relevant'] = [len(df['token_lemma_relevant'][i].split()) for i in range(len(df))]

# Save file

df.to_csv('../DATA/news_125k-ready.csv')

df_tableau = df[['topic', 'topic_name', 'score', 'polarity', 'votes', 'clicks', 'comments', 'num_relevant']]

df_tableau

df_tableau.to_csv('../DATA/news_125k-tableau.csv')
